Create 3
[The program creates a virtual dog and asks the user to
name it. The name is taken and the user is asked whether they would like 
to walk the dog, feed the dog, or clean the dog. Depending on the choice
the three levels (happiness, energy, and hygiene) rise or fall. The program
ends when the user exits the program the 4th option on the main menu.
When the program ends the user is shown the stats of how well they took 
care of their virtual pet based on its happiness level.]